+++ MovingAverage_CUDA_kernel.cuh
@@ -10,7 +10,7 @@ extern "C"
 {
 
     __global__
-    void MovingAverageD(float* I, unsigned int I_size, float* I_S, float* sI, float step);
+    void MovingAverageD(float* I, unsigned int I_size, int* I_S, float* sI, float step);
 
     __global__
     void DivD(unsigned int I_size, float* sI, float step);
